Support unit conversion

+++ app/src/main/java/com/magericx/storagemanipulator/ui/internal_storage/model/InternalStorageInfo.kt
@@ -1,7 +1,7 @@
 package com.magericx.storagemanipulator.ui.internal_storage.model
 
 data class InternalStorageInfo(
-    val availableStorage: Long,
-    val maximumStorage:Long,
+    val availableStorage: Double,
+    val maximumStorage: Double,
     val inUsedCapacityPercent: Double
 )
